+++ Meow/features/data/remote/CatImageRemoteDataSource.swift
@@ -1,0 +1,43 @@
+import Foundation
+
+protocol CatImageRemoteDataSourceProtocol {
+    var urlSession: URLSession { get }
+    var httpRequestHandler: HttpRequestHandler { get }
+    func fetchCatImage() async throws -> CatImage
+}
+
+struct CatImageRemoteDataSource: CatImageRemoteDataSourceProtocol {
+    
+    let urlSession: URLSession
+    let httpRequestHandler: HttpRequestHandler
+    
+    init(urlSession: URLSession, httpRequestHandler: HttpRequestHandler) {
+        self.urlSession = urlSession
+        self.httpRequestHandler = httpRequestHandler
+    }
+    
+    func fetchCatImage() async throws -> CatImage {
+        guard let url: URL = URL(string: APIConfig.baseUrl + APIConfig.getCatImage) else{
+            throw HttpRequestStatus.invalidURLError
+        }
+        
+        do {
+            let response: APIResponse = try await urlSession.data(for: url.createGetURLRequest())
+            
+            let httpRequestStatus: HttpRequestStatus = httpRequestHandler.createRequestStatus(response: response)
+            
+            if httpRequestStatus.hasErrors() {
+                throw httpRequestStatus
+            }
+            
+            guard let catImageApiResponseList: [CatImageApiResponse] = try? JSONDecoder().decode([CatImageApiResponse].self, from: response.data) else {
+                throw HttpRequestStatus.decodingError
+            }
+            
+            return catImageApiResponseList.first?.asCatImage() ?? CatImage()
+            
+        } catch {
+            throw httpRequestHandler.createRequestStatusFrom(error: error)
+        }
+    }
+}
